Put IdentityStore behind interface
To allow end user to swap
